refactor: app.py 코드 경량화 및 모듈화
- app.py 파일 크기 1019줄에서 568줄로 약 44% 감소
- 설정 코드를 config/app_config.py로 분리
- 헬퍼 함수들을 utils/common_helpers.py로 이동
- JWT 콜백 함수들을 utils/jwt_callbacks.py로 분리
- 데이터베이스 초기화 로직을 utils/db_init.py로 분리
- 중복 라우트 제거 및 Blueprint로 통합
- 코드 가독성 및 유지보수성 향상
